Filled continents and countries data, need to fill cities table

+++ src/database_psql/db_setup/create_tables.py
@@ -6,9 +6,9 @@ def setup_continents():
     query = """
     CREATE TABLE IF NOT EXISTS continents (
     name VARCHAR(50) PRIMARY KEY,
-    population INTEGER,
+    population BIGINT,
     world_population REAL,
-    density_per_km INTEGER
+    density_per_km REAL
     );
     """
     cursor.execute(query)
@@ -23,8 +23,8 @@ def setup_countries():
     CREATE TABLE IF NOT EXISTS countries(
     name VARCHAR(50) PRIMARY KEY,
     continent VARCHAR(50) REFERENCES continents (name),
-    population INTEGER,
-    density_per_km2 INTEGER
+    population BIGINT,
+    density_per_km2 REAL
     );
     """
     cursor.execute(query)
@@ -39,7 +39,7 @@ def setup_cities():
     CREATE TABLE IF NOT EXISTS cities(
     name VARCHAR(50) PRIMARY KEY,
     country VARCHAR(50) REFERENCES countries (name),
-    population INTEGER
+    population BIGINT
     );
     """
     cursor.execute(query)
